refactor(logic): widen id method parameters int -> long
Follow-up to the response-field widening. ESI types path, query, and body
integer parameters as int64 as well, so leaving the method parameters at
int forced a downcast on the round trip (a long id read from one call
could not be passed into the next).

- Every id-shaped parameter in Logic is now long: scalar `int foo_id`,
  `int[] foo_ids`, `List<int>`, `int?`. Includes wallet `division`
  (int64 in the spec, no bound).
- Passing an int argument still compiles (implicit widening); callers only
  change where they built a `List<int>` / `int[]` to pass in.
- Integration-test fixture ids and the Universe.Names call updated to long;
  the temporary (int) casts removed.

SpecCheck unchanged (RESULT: clean, 0 errors) - it only checks response
bodies. MIGRATION.md / CHANGELOG.md section updated to cover parameters.

+++ ESI.NET/Logic/AllianceLogic.cs
@@ -28,7 +28,7 @@ namespace ESI.NET.Logic
         /// </summary>
         /// <param name="allianceId"></param>
         /// <returns></returns>
-        public async Task<EsiResponse<Alliance>> Information(int alliance_id, EsiCallOptions options = null)
+        public async Task<EsiResponse<Alliance>> Information(long alliance_id, EsiCallOptions options = null)
             => await Execute<Alliance>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/alliances/{alliance_id}/",
                 replacements: new Dictionary<string, string>()
                 {
@@ -42,7 +42,7 @@ namespace ESI.NET.Logic
         /// </summary>
         /// <param name="alliance_id"></param>
         /// <returns></returns>
-        public async Task<EsiResponse<long[]>> Corporations(int alliance_id, EsiCallOptions options = null)
+        public async Task<EsiResponse<long[]>> Corporations(long alliance_id, EsiCallOptions options = null)
             => await Execute<long[]>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/alliances/{alliance_id}/corporations/",
                 replacements: new Dictionary<string, string>()
                 {
@@ -56,7 +56,7 @@ namespace ESI.NET.Logic
         /// </summary>
         /// <param name="alliance_id"></param>
         /// <returns></returns>
-        public async Task<EsiResponse<Images>> Icons(int alliance_id, EsiCallOptions options = null)
+        public async Task<EsiResponse<Images>> Icons(long alliance_id, EsiCallOptions options = null)
             => await Execute<Images>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/alliances/{alliance_id}/icons/",
                 replacements: new Dictionary<string, string>()
                 {
